+++ crates/teloxide-core/src/adaptors/throttle/requester_impl.rs
@@ -204,7 +204,8 @@ where
         get_game_high_scores,
         get_available_gifts,
         send_gift,
-        set_user_emoji_status
+        set_user_emoji_status,
+        save_prepared_inline_message
         => fid, ftyid
     }
 }