+++ crates/teloxide-core/CHANGELOG.md
@@ -78,7 +78,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for TBA 8.0 ()
   - Add `Gift`, `Gifts` and `PreparedInlineMessage`  structs
-  - Add `get_available_gifts`, `send_gift` and `set_user_emoji_status` TBA methods
+  - Add `get_available_gifts`, `send_gift`, `set_user_emoji_status` and `save_prepared_inline_message` TBA methods
   - Add `Gift` field to `TransactionPartnerUser`.
 
 [pr1157]: https://github.com/teloxide/teloxide/pull/1157